+++ src/Zmsdb/Helper/SendNotificationReminder.php
@@ -4,46 +4,89 @@ namespace BO\Zmsdb\Helper;
 
 class SendNotificationReminder
 {
-    protected $processList;
-
     protected $datetime;
+
+    protected $reminderInSeconds;
 
     protected $verbose = false;
 
-    public function __construct($verbose = false, \DateTimeInterface $now = null)
+    protected $limit = 10000;
+
+    protected $loopCount = 500;
+
+    protected $count = 0;
+
+    public function __construct(\DateTimeInterface $now, $verbose = false)
     {
-        if (!$now) {
-            $now = new \DateTimeImmutable();
-        }
         $this->dateTime = $now;
         if ($verbose) {
-            error_log("INFO: Send notification reminder dependent on lead time");
             $this->verbose = true;
+            $this->log("\nINFO: Send notification reminder dependent on lead time");
         }
+    }
 
-        $this->processList = (new \BO\Zmsdb\Process)->readNotificationReminderProcessList($this->dateTime, 10000, 2);
+    protected function log($message)
+    {
+        if ($this->verbose) {
+            error_log($message);
+        }
+    }
+
+    public function getCount()
+    {
+        return $this->count;
+    }
+
+    public function setLimit($limit)
+    {
+        $this->limit = $limit;
+    }
+
+    public function setLoopCount($loopCount)
+    {
+        $this->loopCount = $loopCount;
     }
 
     public function startProcessing($commit)
     {
-        foreach ($this->processList as $process) {
-            if ($this->verbose) {
-                error_log("INFO: Processing $process");
-            }
-            if ($commit) {
-                $notification = $this->writeNotification($process);
-                if ($notification) {
-                    if ($this->updateProcessRemoveReminder($process) && $this->verbose) {
-                        error_log("INFO: Updated $process->id - reminder timestamp removed");
-                    }
-                } else {
-                    error_log("WARNING: Notification for $process->id not possible - no telephone or not enabled");
-                }
-            }
-        }
+        $this->writeNotificationReminderList($commit);
+        $this->log("\nSUMMARY: Sent notification reminder: ".$this->count);
     }
 
-    protected function writeNotification(\BO\Zmsentities\Process $process)
+    protected function writeNotificationReminderList($commit)
+    {
+        $count = $this->writeByCallback($commit, function ($limit, $offset) {
+            $processList = (new \BO\Zmsdb\Process)->readNotificationReminderProcessList(
+                $this->dateTime,
+                $limit,
+                $offset,
+                2
+            );
+            return $processList;
+        });
+        $this->count += $count;
+    }
+
+    protected function writeByCallback($commit, \Closure $callback)
+    {
+        $processCount = 0;
+        $startposition = 0;
+        while ($processCount < $this->limit) {
+            $processList = $callback($this->loopCount, $startposition);
+            if (0 == $processList->count()) {
+                break;
+            }
+            foreach ($processList as $process) {
+                if (!$this->writeReminder($process, $commit, $processCount)) {
+                    $startposition++;
+                }
+                $processCount++;
+            }
+        }
+        return $processCount;
+    }
+
+    protected function writeReminder(\BO\Zmsentities\Process $process, $commit, $processCount)
     {
         $notification = null;
         $department = (new \BO\Zmsdb\Department)->readByScopeId($process->getScopeId(), 2);
@@ -52,15 +95,29 @@ class SendNotificationReminder
         ) {
             $config = (new \BO\Zmsdb\Config)->readEntity();
             $entity = (new \BO\Zmsentities\Notification)->toResolvedEntity($process, $config, $department, 'reminder');
-            $notification = (new \BO\Zmsdb\Notification)->writeInQueue($entity, $this->dateTime);
+            if ($commit) {
+                $notification = (new \BO\Zmsdb\Notification)->writeInQueue($entity, $this->dateTime);
+                $this->log(
+                    "\nINFO: $processCount. Write reminder notification in queue with ID ". $notification->getId() . " for process id ". $process->getId()
+                );
+                $this->deleteReminderTimestamp($process, $notification, $processCount, $commit);
+            }
         }
         return $notification;
     }
 
-    protected function updateProcessRemoveReminder(\BO\Zmsentities\Process $process)
-    {
-        $process->reminderTimestamp = 0;
-        $process = (new \BO\Zmsdb\Process)->updateEntity($process, $this->dateTime);
-        return $process;
+    protected function deleteReminderTimestamp(
+        \BO\Zmsentities\Process $process,
+        $notification,
+        $processCount,
+        $commit
+    ) {
+        if ($notification) {
+            $process->reminderTimestamp = 0;
+            $process = (new \BO\Zmsdb\Process)->updateEntity($process, $this->dateTime);
+            $this->log("INFO: $processCount. Updated $process->id - reminder timestamp removed");
+        } else {
+            $this->log("\nWARNING: Notification for $process->id not possible - no telephone or not enabled");
+        }
     }
 }